refactoring tool-lib boilerplate into base file
